add change tienda modal and admin login update

+++ PointOfSale.Business/Contracts/IAfipService.cs
@@ -12,7 +12,7 @@ namespace PointOfSale.Business.Contracts
 {
     public interface IAfipService
     {
-        Task<FacturaEmitida> Facturar(Sale sale_created, int? nroDocumento, int? idCliente, string registrationUser);
+        Task<FacturaEmitida> Facturar(Sale sale_created, long? nroDocumento, int? idCliente, string registrationUser);
 
         Task<List<FacturaEmitida>> GetAll(int idTienda);
